+++ src/laraifox/minecraft/world/Stack.java
@@ -38,7 +38,7 @@ public class Stack {
 		return chunks[y / Chunk.CHUNK_SIZE].getBlock(x, (int) (y % Chunk.CHUNK_SIZE), z);
 	}
 
-	public void setBlock(World world, short id, int x, int y, int z) {
+	public void setBlock(World world, int id, int x, int y, int z) {
 		chunks[y / Chunk.CHUNK_SIZE].setBlock(world, id, x, (int) (y % Chunk.CHUNK_SIZE), z);
 	}
